. list vector bubbles tests
