+++ src/o/list.js
@@ -1,16 +1,16 @@
-const List = require("./list");
+const LinkedList = require("./linked_list");
 
-let emptyBubbles;
+let emptyList;
 
-class Bubbles extends List {
+class List extends LinkedList {
 
-  static listOpenChar = "(";
-  static listCloseChar = ")";
+  static linkedListOpenChar = "(";
+  static linkedListCloseChar = ")";
 
-  static get emptyBubbles() { return emptyBubbles; }
+  static get emptyList() { return emptyList; }
 
-  // Create a bubbles.
-  constructor(o, oo=emptyBubbles) {
+  // Create a list.
+  constructor(o, oo=emptyList) {
     super(o, oo);
   }
 
@@ -18,17 +18,17 @@ class Bubbles extends List {
     return _make(elements);
   }
 
-  static blow(...bubbles) {
-    return _blow(bubbles);
+  static blow(...list) {
+    return _blow(list);
   }
 
   static from(arrayLike, mapFn, thisArg) {
     let array = Array.from(arrayLike, mapFn, thisArg);
-    return Bubbles.make(...array);
+    return List.make(...array);
   }
 
-  static push(list, element) { // oo -> list, o -> element
-    return new Bubbles(element, list); // o -> element, oo -> list
+  static push(linkedList, element) { // oo -> linkedList, o -> element
+    return new List(element, linkedList); // o -> element, oo -> linkedList
   }
 
   static invert(o) {
@@ -44,7 +44,7 @@ class Bubbles extends List {
 
   static map(o, fn) {
     if (!o.isEmpty)
-      return new Bubbles(fn(o.o),
+      return new List(fn(o.o),
         map(o.oo, fn));
   }
 
@@ -64,15 +64,15 @@ class Bubbles extends List {
 
     // oo (accumulatedString), o (formattedElement)
     let join = (accumulatedString, formattedElement) => {
-      return accumulatedString + " " + formattedElement; // Original was `oo + " " + o`, if list is reversed, this should be `formattedElement + " " + accumulatedString`
+      return accumulatedString + " " + formattedElement; // Original was `oo + " " + o`, if linkedList is reversed, this should be `formattedElement + " " + accumulatedString`
                                                         // However, looking at LynktLyst.toString, it was `o + " " + oo`.
                                                         // Let's keep the original logic of prepending: `formattedElement + " " + accumulatedString`
-                                                        // if the reduce iterates head to tail and wants to build a reversed string to match (e.g. LIFO bubbles -> string)
+                                                        // if the reduce iterates head to tail and wants to build a reversed string to match (e.g. LIFO list -> string)
                                                         // Or, if reduce iterates head to tail and we want natural order string, it should be `accumulatedString + " " + formattedElement`
-                                                        // Given `LynktLyst.toString` also had `o + " " + oo` and it works to produce `(1 2 3)`, this implies reduce iterates from tail (or list is inverted before stringification).
-                                                        // For now, I keep the parameter names and the original logic: `formattedElement + " " + accumulatedString` assuming it's correct for the list's iteration order in reduce.
+                                                        // Given `LynktLyst.toString` also had `o + " " + oo` and it works to produce `(1 2 3)`, this implies reduce iterates from tail (or linkedList is inverted before stringification).
+                                                        // For now, I keep the parameter names and the original logic: `formattedElement + " " + accumulatedString` assuming it's correct for the linkedList's iteration order in reduce.
                                                         // The original code `oo + " " + o` means `accumulatedString + " " + formattedElement`.
-                                                        // If `reduce` processes from head (e.g. 1, then 2, then 3 for list (1 2 3) ):
+                                                        // If `reduce` processes from head (e.g. 1, then 2, then 3 for linkedList (1 2 3) ):
                                                         // Iter 1: memo="", current=1 -> memo=" 1"
                                                         // Iter 2: memo=" 1", current=2 -> memo=" 1 2"
                                                         // This seems more standard. Let's use `accumulatedString + " " + formattedElement`.
@@ -84,41 +84,41 @@ class Bubbles extends List {
          + ")";
   }
 
-  toString() { return Bubbles.toString(this); } // Ensure static toString is called for consistency
-  push(element) { return Bubbles.push(this, element); } // o -> element, ensure static push
+  toString() { return List.toString(this); } // Ensure static toString is called for consistency
+  push(element) { return List.push(this, element); } // o -> element, ensure static push
 
 }
 
 
 // These are static methods, ensure they
-// are used as .map, Bubbles.push etc.
+// are used as .map, List.push etc.
 // if needed inside instance methods, or
 // this is fine if they are standalone
 // pure functions from LynktLyst.
 const { map, push, reduce, toString,
-        pop, peek, make } = Bubbles;
+        pop, peek, make } = List;
 
 
-function _make(elementsArray, currentList=emptyBubbles) {
+function _make(elementsArray, currentLinkedList=emptyList) {
   if (elementsArray.length < 1)
-    return currentList;
+    return currentLinkedList;
   return _make(elementsArray,
-    new Bubbles(elementsArray.pop(),
-      currentList));
+    new List(elementsArray.pop(),
+      currentLinkedList));
 }
 
-function _blow(elementsArray, currentList=emptyBubbles) {
+function _blow(elementsArray, currentLinkedList=emptyList) {
   if (elementsArray.length < 1)
-    return currentList;
+    return currentLinkedList;
   return _blow(elementsArray,
-    new Bubbles(elementsArray.pop(),
-      currentList));
+    new List(elementsArray.pop(),
+      currentLinkedList));
 }
 
-class EmptyBubbles {
+class EmptyList {
   get isEmpty() { return true; }
 }
 
-emptyBubbles = new EmptyBubbles()
+emptyList = new EmptyList()
 
-module.exports = Bubbles;
+module.exports = List;
